+++ server/src/hover.ts
@@ -1,0 +1,8 @@
+import { MarkupContent, Position } from "vscode-languageserver";
+import { VNSFile } from "./types";
+
+export const checkHoverInfo = (file: VNSFile, pos: Position): MarkupContent => {
+    let result: MarkupContent = null;
+
+    return result;
+};