+++ server/src/server.ts
@@ -10,20 +10,32 @@ import {
     TextDocumentPositionParams,
     TextDocumentSyncKind,
     InitializeResult,
-    DocumentColorRequest
+    DocumentColorRequest,
+    HandlerResult,
+    Hover
 } from 'vscode-languageserver/node';
 
 import { TextDocument } from 'vscode-languageserver-textdocument';
 
 import { inspectVNS } from './inspector';
 
+import { VNSFile } from "./types";
+import { checkHoverInfo } from './hover';
+
 let connection = createConnection(ProposedFeatures.all);
 
 let documents = new TextDocuments(TextDocument);
 
+const documentElements: Map<string, VNSFile> = new Map();
+
 // adjust paramters for initializing
 connection.onInitialize((params) => {
-    return { capabilities: { textDocumentSync: TextDocumentSyncKind.Full } };
+    return {
+        capabilities: {
+            textDocumentSync: TextDocumentSyncKind.Full,
+            hoverProvider: true
+        }
+    };
 });
 
 // register configuration notification channel to listen
@@ -54,14 +66,35 @@ documents.onDidClose((change) => {
     });
 });
 
+// show hover infos
+connection.onHover((params, token, workDoneProgress, resultProgress?): HandlerResult<Hover, null> => {
+    let result: Hover = null;
+
+    // TODO: query for hover infos
+    let file = params.textDocument;
+    if (documentElements.has(file.uri)) {
+        result = { contents: checkHoverInfo(documentElements.get(file.uri), params.position) };
+        if (result.contents === null) {
+            result = null;
+        }
+    }
+
+    return result;
+});
+
 // inspect files asynchronically
 const inspect = async (file: TextDocument) => {
-    const errors = inspectVNS(file);
+    const inspectResult = inspectVNS(file);
 
     connection.sendDiagnostics({
         uri: file.uri,
-        diagnostics: errors
+        diagnostics: inspectResult.errors
     });
+
+    if (documentElements.has(file.uri)) {
+        documentElements.delete(file.uri);
+    }
+    documentElements.set(file.uri, inspectResult.file);
 };
 
 // register listeners